Move `IHttpResponseStreamWriterFactory` too
nits: take VS suggestions for changed files
