+++ src/Microsoft.AspNetCore.Mvc.Core/Infrastructure/IHttpResponseStreamWriterFactory.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using System.Text;
 
-namespace Microsoft.AspNetCore.Mvc.Internal
+namespace Microsoft.AspNetCore.Mvc.Infrastructure
 {
     /// <summary>
     /// Creates <see cref="TextWriter"/> instances for writing to <see cref="Http.HttpResponse.Body"/>.
